tech-debt: renames repo
