--- src/aind_labtracks_service/__init__.py
@@ -1,3 +1,0 @@
-"""REST service to retrieve data from labtracks.
-"""
-__version__ = "0.0.0"
