+++ src/aind_metadata_service/__init__.py
@@ -1,0 +1,3 @@
+"""REST service to retrieve metadata from databases.
+"""
+__version__ = "0.0.0"
